[profiler] provide torch.profiler._utils._init_for_cuda_graphs() as a workaround
There are known issues with profiling cuda graphs - particularly, if you create a cuda graph before the first use of the profiler, and then run that cuda graph during profiling.

One workaround is to add `with profile(): pass` before creating the cuda graph that you want to profile later.

For convenience, we provide this function to use the workaround. This also adads a test for this workaround, to ensure that it continues working.

ghstack-source-id: 41b6e6c119d7034257a02284338d24f069fb1306
Pull Request resolved: https://github.com/pytorch/pytorch/pull/100441