dynamic shapes mismatch errors (#132982)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/132982

When PyTree detects a structural mismatch between inputs and dynamic shapes, the error messages are quite horrible. This PR fixes these error messages by adding, for each kind of error, the path to the point where the error happens and an actionable reason for the error.

Test Plan: added test with several cases

Reviewed By: yushangdi

Differential Revision: D60956976