[ROCm][TunableOp] improve selection criteria
Previously, the solution with the fastest average time was selected.
Now, the max time is also considered as a worst-case. The best solutions
must be fastest on average as well as have a better worst-case.

The criteria for skipping slow solutions was increased to 3x slower
than the current fastest, up from 2x.

The Default solution is now retried as the final solution in addition to
being tried first. There seemed to be a penalty against the first
solution attempted.