Generate unique id for tensor storage object by observing the week pointer of tensor storage object (#154859)
Summary:

PyTorch execution trace records tensor storage data in the trace. The tensor storage data includes storage id, offset, number of elements, and number of byte for each element. PARAM et-replay uses this information to allocate/free the tensors.
However, the current implementation of generating tensor storage id does not guarantee it is unique. ExecutionTraceObserver maintains a lookup table to map the memory address of the tensor storage object to an unique id. If a new memory address is found, it will be put into that hash table and associate it to a new id.
This implementation does not guarantee the storage object id is unique since the memory that the address points to may be released and then re-allocated to a different tensor storage object.
This DIFF is to observe the week pointer of tensor storage object in ET. When a tensor storage is deleted, its week pointer will expire. ET saves a map between raw data pointer to the week pointer of the tensor storage object. If the memory address got reused, the week pointer in the map will expire, then the new ID will be generated for it, and the map will be updated with the week pointer to the new tensor storage object.

Test Plan:
buck2 run mode/opt caffe2/test:test_profiler_cuda -- profiler.test_execution_trace.TestExecutionTraceCUDA

Rollback Plan:

Reviewed By: eellison

Differential Revision: D75749065