[Inductor] Fix epilogue fusion decision with 1 Triton caller as choice (#156500)
Summary:


In the current scheduler logic, if a template buffer is only a Triton template, which can result from only 1 Triton choice in the autotuning, the fusion won't be benchmarked.

This can lead to an edge case in which a Triton GEMM template from the autotune lookup table can have a problematic fusion, leading to shared memory requirements above the hardware limit. `(256, 128, 64, 4, 8, 8)` is such a config, where we have seen fusion with a `to(torch.float32)` can lead to this issue, `out of resource: shared memory, Required: 264224, Hardware limit: 232448`. We benchmark the fusion for this case to ensure it's safe.

Test Plan:
`buck run mode/opt fbcode//caffe2/test/inductor:max_autotune 2>&1 | tee ~/test.log`

Rollback Plan:

Reviewed By: jansel

Differential Revision: D76904773