[dtensor] fix incorrect norm calculation for Partial DTensors
The sharding strategies for `aten.linalg_vector_norm` and the optimized
`aten._foreach_norm.Scalar` incorrectly assumes the norm operation is
always "reduction linear" with respect to its inputs. This bug causes
the norm to be computed on local, incomplete data for DTensors with a
`Partial(sum)` placement, leading to an inflated result (a sum of norms,
rather than the correct norm of the sum).

This fix simply forces `reduction_linear=False` for partial placements.