Add waitcounters for torch.compile subprocess pool
Summary:
This ads waitcounter for whether or not the pool is running, as well as if we
are running jobs.

This also ads waitcounters for the first job within a pool.

Test Plan:
We see valid waitcoutners when we rebase this on a job we are debugging -
https://fburl.com/canvas/zm40n4e2

As well as indicating shutdown with async_compile_quiesce enabled
https://fburl.com/canvas/uykbkoab

Differential Revision: D83705931