torch.compile: Increase subprocess parent death check interval to lower cpu
Summary:
This check is a good idea (we could potentially do it with prctl). However
we're seeing elevated rates of cpu usage in idle worker threads. This causes issues on production jobs, causing a large amount of spikeness in qps.

Test Plan:
Tested on a prod job with caches force disabled via
TORCH_COMPILE_FORCE_DISABLE_CACHES=1

Baseline
 {F1982479204}

With this diff -
 {F1982479208}

Differential Revision: D83803302